refactor: split out code into smaller files
